onlinegdb--empirically_determine_max_thread_stack_size_GS_version.c: fix this program by using the alloca() buf
The program didn't work until I added the usage of the buffer allocated by `alloca()` back
in. It seems without actually writing to this allocated memory, perhaps the compiler
was optimizing it out after-all!

I kept `volatile` in too just in case some systems might try to optimize out
the write to the buffer without it.