strncmpci.c: remove `goto`
Remove `goto` from the main function, but copy the function and retain
it in a second version of the function.

This example code has super simple error handling, and goto isn't
required. goto is a very valid, good, and valuable tool to make clean,
safe and readable error-handling code in C, _especially_ for the really
complicated cases where you are initializing a ton of hardware and need
to to different amounts of cleanup, and in different ways, if something
fails during this long initialization process.

Nevertheless, some people may be turned off by it because they don't
understand it and they've been misinformed about it their whole lives
and told it's evil. For this super simple case, there's no sense in
having to try to convince people, so just change it.

Note to anyone who's curious: here's some of the valid justification
I've written about for the virtues of goto. See my little blurb at the
bottom of my answer here, titled

------------------------------------------------------------------------
**"Additional reading and justification
for valid usage of `goto` in error handling for professional code":**

https://stackoverflow.com/a/54488289/4561887.
------------------------------------------------------------------------