Rename c/utilities.cpp --> c/utilities.c (oops!)
This is actually a C program. I didn't mean to give it a C++ extension.

Maybe sometime I'll move it into the C++ folder though...? Nah...
instead, when you need to add templates or C++ features or something,
just add a cpp/utilities_cpp.h and cpp/utilities_cpp.cpp or something
instead.

**If you do that, be sure to include this C utilities.h in that C++
version, using `extern "C" {}`, so you can just include the C one in C++
as well!**

See: https://stackoverflow.com/questions/1041866/what-is-the-effect-of-extern-c-in-c/30526795#30526795