python/git_get_short_hash.py: do bug fix!
I don't think the `'"$(git status --porcelain)"'` part in the
`subprocess.run()` call was getting evaluated before, so `test` was
erroneously always indicating that `git status` was dirty!

This fixes it to work properly in Python, and only indicate that the
`git status` is dirty when it really is dirty.